+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 _Not released yet_
 
 - Breaking Change: Azure Data Explorer plugin now requires Grafana 8.0+ to run.
-- Breaking Change: obo_latency_seconds metric was removed
+- Breaking Change: obo_latency_seconds metric was removed.
+- Bugfix: Included new Kusto query editor. **NOTE**: This new editor will be only available if used with Grafana 8.5 or later. Fixes #325.
 - Bugfix: Filter dynamic columns from Where/Aggregate/Group by clauses to prevent syntax errors.
-- Internal: Client secret authentication via Grafana Azure SDK
-- Internal: OBO authentication via MSAL for Go
+- Bugfix: Add logical operators for timespan types in the query builder.
+- Internal: Client secret authentication via Grafana Azure SDK.
+- Internal: OBO authentication via MSAL for Go.
 
 ## [3.7.1]